Docs/scripts/requirements.txt
# ============================================
# DEPENDÊNCIAS DO SISTEMA DE GESTÃO DE CONHECIMENTO
# ============================================

# Core
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0

# OpenAI & LLMs
openai>=1.52.0
tiktoken>=0.7.0

# NLP & Text Processing
spacy>=3.7.0
sentence-transformers>=2.7.0
nltk>=3.8.1

# Similarity & Clustering
scikit-learn>=1.5.0
umap-learn>=0.5.5
hdbscan>=0.8.33
numpy>=1.26.0
pandas>=2.2.0

# Graph & Visualization
networkx>=3.3
pyvis>=0.3.2
plotly>=5.22.0
matplotlib>=3.9.0

# Markdown & Obsidian
python-frontmatter>=1.1.0
markdown>=3.6
beautifulsoup4>=4.12.3

# File Processing
chardet>=5.2.0
pathlib>=1.0.1

# Scheduling
schedule>=1.2.0
watchdog>=4.0.0

# Utils
python-dateutil>=2.9.0
colorama>=0.4.6
